move text builder part into another file.
